+++ tests/acceptance/steps/when.py
@@ -1,6 +1,7 @@
 """
 'When' step implementations for acceptance tests.  Powered by behave.
 """
+
 from pathlib import Path