+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <https://semver.org>`__.
 
+0.8.0 (unreleased)
+------------------
+
+- Drop direnv support (.envrc configuration file)
+- Drop PHP project support (Symfony and TYPO3 Web frameworks)
+
 0.7.0 (2022-02-19)
 ------------------